WB-34: move viewmodels up a level — lib/logic/viewmodels → lib/viewmodels
The extra `logic/` segment had no meaning once viewmodels
became a first-class concept. Flattens the path and drops the
ExampleViewModel references in the README to match the new
require paths (`require("viewmodels/Foo")`).
